Move BlockByHash -> ByHash
